separate unit and json tests
