+++ xrpl4j-core/src/main/java/org/xrpl/xrpl4j/model/transactions/DepositPreAuth.java
@@ -87,18 +87,18 @@ public interface DepositPreAuth extends Transaction {
   /**
    * The {@link CredentialWrapper}'s to preauthorize.
    *
-   * @return An {@link Optional} list of type {@link CredentialWrapper} to preauthorize.
+   * @return A list of type {@link CredentialWrapper} to preauthorize.
    */
   @JsonProperty("AuthorizeCredentials")
-  Optional<List<CredentialWrapper>> authorizeCredentials();
+  List<CredentialWrapper> authorizeCredentials();
 
   /**
    * The {@link CredentialWrapper}'s whose preauthorization should be revoked.
    *
-   * @return An {@link Optional} list of type {@link CredentialWrapper} to unauthorize.
+   * @return A list of type {@link CredentialWrapper} to unauthorize.
    */
   @JsonProperty("UnauthorizeCredentials")
-  Optional<List<CredentialWrapper>> unauthorizeCredentials();
+  List<CredentialWrapper> unauthorizeCredentials();
 
   /**
    * Validate that exactly one of {@link DepositPreAuth#authorize()} or {@link DepositPreAuth#unauthorize()} or
@@ -113,10 +113,10 @@ public interface DepositPreAuth extends Transaction {
     if (unauthorize().isPresent()) {
       fieldsPresent++;
     }
-    if (authorizeCredentials().isPresent()) {
+    if (!authorizeCredentials().isEmpty()) {
       fieldsPresent++;
     }
-    if (unauthorizeCredentials().isPresent()) {
+    if (!unauthorizeCredentials().isEmpty()) {
       fieldsPresent++;
     }
 
@@ -130,19 +130,19 @@ public interface DepositPreAuth extends Transaction {
    */
   @Value.Check
   default void validateCredentialList() {
-    authorizeCredentials().ifPresent(creds ->
+    if (!authorizeCredentials().isEmpty()) {
       Preconditions.checkArgument(
-        !creds.isEmpty() && creds.size() <= 8,
-        "AuthorizeCredentials shouldn't be empty and must have less than or equal to 8 credentials."
-      )
-    );
+        authorizeCredentials().size() <= 8,
+        "AuthorizeCredentials should have less than or equal to 8 credentials."
+      );
+    }
 
-    unauthorizeCredentials().ifPresent(creds ->
+    if (!unauthorizeCredentials().isEmpty()) {
       Preconditions.checkArgument(
-        !creds.isEmpty() && creds.size() <= 8,
-        "UnauthorizeCredentials shouldn't be empty and must have less than or equal to 8 credentials."
-      )
-    );
+        unauthorizeCredentials().size() <= 8,
+        "UnauthorizeCredentials should have less than or equal to 8 credentials."
+      );
+    }
   }
 
   /**
@@ -151,14 +151,20 @@ public interface DepositPreAuth extends Transaction {
    */
   @Value.Check
   default void validateForUniqueValues() {
-    authorizeCredentials().ifPresent(creds -> Preconditions.checkArgument(
-      new HashSet<>(creds).size() == creds.size(),
-      "AuthorizeCredentials should have unique credentials."
-    ));
+    if (!authorizeCredentials().isEmpty()) {
+      final List<CredentialWrapper> credentials = authorizeCredentials();
+      Preconditions.checkArgument(
+        new HashSet<>(credentials).size() == credentials.size(),
+        "AuthorizeCredentials should have unique credentials."
+      );
+    }
 
-    unauthorizeCredentials().ifPresent(creds -> Preconditions.checkArgument(
-      new HashSet<>(creds).size() == creds.size(),
-      "UnauthorizeCredentials should have unique credentials."
-    ));
+    if (!unauthorizeCredentials().isEmpty()) {
+      final List<CredentialWrapper> credentials = unauthorizeCredentials();
+      Preconditions.checkArgument(
+        new HashSet<>(credentials).size() == credentials.size(),
+        "UnauthorizeCredentials should have unique credentials."
+      );
+    }
   }
 }
