+++ xrpl4j-core/src/main/java/org/xrpl/xrpl4j/model/transactions/PaymentChannelClaim.java
@@ -136,20 +136,22 @@ public interface PaymentChannelClaim extends Transaction {
    * Set of Credentials to authorize a deposit made by this transaction. Each member of the array must be the ledger
    * entry ID of a Credential entry in the ledger.
    *
-   * @return An {@link Optional} of type {@link Hash256}.
+   * @return A list of type {@link Hash256}.
    */
   @JsonProperty("CredentialIDs")
-  Optional<List<Hash256>> credentialIds();
+  List<Hash256> credentialIds();
 
   /**
    * Validate {@link PaymentChannelClaim#credentialIds} has less than or equal to 8 credentials.
    */
   @Value.Check
   default void validateCredentialIdsLength() {
-    credentialIds().ifPresent(credentialIds -> Preconditions.checkArgument(
-      !credentialIds.isEmpty() && credentialIds.size() <= 8,
-      "credentialIds shouldn't be empty and must have less than or equal to 8 items."
-    ));
+    if (!credentialIds().isEmpty()) {
+      Preconditions.checkArgument(
+        credentialIds().size() <= 8,
+        "CredentialIDs should have less than or equal to 8 items."
+      );
+    }
   }
 
   /**
@@ -157,9 +159,11 @@ public interface PaymentChannelClaim extends Transaction {
    */
   @Value.Check
   default void validateUniqueCredentialIds() {
-    credentialIds().ifPresent(credentialIds -> Preconditions.checkArgument(
-      new HashSet<>(credentialIds).size() == credentialIds.size(),
-      "credentialIds should have unique values."
-    ));
+    if (!credentialIds().isEmpty()) {
+      Preconditions.checkArgument(
+        new HashSet<>(credentialIds()).size() == credentialIds().size(),
+        "CredentialIDs should have unique values."
+      );
+    }
   }
 }
