+++ xrpl4j-core/src/test/java/org/xrpl/xrpl4j/model/transactions/PaymentChannelCreateTest.java
@@ -21,18 +21,15 @@ package org.xrpl.xrpl4j.model.transactions;
  */
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import com.google.common.primitives.UnsignedInteger;
 import com.google.common.primitives.UnsignedLong;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-
-public class PaymentChannelTests {
+/**
+ * Unit tests for {@link PaymentChannelCreate}.
+ */
+public class PaymentChannelCreateTest {
 
   @Test
   public void createWithoutCancelAfterOrDestinationTag() {
@@ -102,69 +99,4 @@ public class PaymentChannelTests {
     assertThat(create.flags().isEmpty()).isTrue();
     assertThat(create.cancelAfter()).isPresent().get().isEqualTo(UnsignedLong.valueOf(533171558));
   }
-
-  @Test
-  public void testMoreThanEightCredentialIds() {
-    List<Hash256> moreThanEight = IntStream.range(0, 9)
-      .mapToObj(i ->
-        Hash256.of("7C221D901192C74AA7AC60786B1B01A88E922BE267E5B5B4FA64D214C5067FF" + i))
-      .collect(Collectors.toList());
-
-    assertThrows(
-      IllegalArgumentException.class,
-      () -> PaymentChannelClaim.builder()
-        .account(Address.of("rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn"))
-        .fee(XrpCurrencyAmount.ofDrops(100))
-        .sequence(UnsignedInteger.ONE)
-        .amount(XrpCurrencyAmount.ofDrops(10000))
-        .publicKey("32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A")
-        .credentialIds(moreThanEight)
-        .channel(Hash256.of("7C221D901192C74AA7AC60786B1B01A88E922BE267E5B5B4FA64D214C5067FF1"))
-        .build(),
-      "credentialIds shouldn't be empty and must have less than or equal to 8 items."
-    );
-
-  }
-
-  @Test
-  public void testEmptyCredentialIds() {
-    assertThrows(
-      IllegalArgumentException.class,
-      () -> PaymentChannelClaim.builder()
-        .account(Address.of("rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn"))
-        .fee(XrpCurrencyAmount.ofDrops(100))
-        .sequence(UnsignedInteger.ONE)
-        .amount(XrpCurrencyAmount.ofDrops(10000))
-        .publicKey("32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A")
-        .credentialIds(new ArrayList<>())
-        .channel(Hash256.of("7C221D901192C74AA7AC60786B1B01A88E922BE267E5B5B4FA64D214C5067FF1"))
-        .build(),
-      "credentialIds shouldn't be empty and must have less than or equal to 8 items."
-    );
-  }
-
-  @Test
-  public void testDuplicateCredentialIds() {
-    List<Hash256> randomIds = IntStream.range(0, 8)
-      .mapToObj(i ->
-        Hash256.of("7C221D901192C74AA7AC60786B1B01A88E922BE267E5B5B4FA64D214C5067FF" + i))
-      .collect(Collectors.toList());
-
-    randomIds.set(1, randomIds.get(0));
-
-    assertThrows(
-      IllegalArgumentException.class,
-      () -> PaymentChannelClaim.builder()
-        .account(Address.of("rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn"))
-        .fee(XrpCurrencyAmount.ofDrops(100))
-        .sequence(UnsignedInteger.ONE)
-        .amount(XrpCurrencyAmount.ofDrops(10000))
-        .publicKey("32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A")
-        .credentialIds(randomIds)
-        .channel(Hash256.of("7C221D901192C74AA7AC60786B1B01A88E922BE267E5B5B4FA64D214C5067FF1"))
-        .build(),
-      "credentialIds should have unique values."
-    );
-  }
-
 }
