--- xrpl4j-core/src/test/java/org/xrpl/xrpl4j/model/transactions/json/PaymentChannelJsonTests.java
@@ -1,433 +1,0 @@
-package org.xrpl.xrpl4j.model.transactions.json;
-
-/*-
- * ========================LICENSE_START=================================
- * xrpl4j :: model
- * %%
- * Copyright (C) 2020 - 2022 XRPL Foundation and its contributors
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * =========================LICENSE_END==================================
- */
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.google.common.primitives.UnsignedInteger;
-import com.google.common.primitives.UnsignedLong;
-import org.json.JSONException;
-import org.junit.jupiter.api.Test;
-import org.xrpl.xrpl4j.crypto.keys.PublicKey;
-import org.xrpl.xrpl4j.model.AbstractJsonTest;
-import org.xrpl.xrpl4j.model.flags.PaymentChannelClaimFlags;
-import org.xrpl.xrpl4j.model.flags.TransactionFlags;
-import org.xrpl.xrpl4j.model.transactions.Address;
-import org.xrpl.xrpl4j.model.transactions.Hash256;
-import org.xrpl.xrpl4j.model.transactions.NetworkId;
-import org.xrpl.xrpl4j.model.transactions.PaymentChannelClaim;
-import org.xrpl.xrpl4j.model.transactions.PaymentChannelCreate;
-import org.xrpl.xrpl4j.model.transactions.PaymentChannelFund;
-import org.xrpl.xrpl4j.model.transactions.XrpCurrencyAmount;
-
-import java.util.Collections;
-
-public class PaymentChannelJsonTests extends AbstractJsonTest {
-
-  @Test
-  public void testPaymentChannelCreateJson() throws JsonProcessingException, JSONException {
-    PaymentChannelCreate create = PaymentChannelCreate.builder()
-      .account(Address.of("rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn"))
-      .sourceTag(UnsignedInteger.valueOf(11747))
-      .fee(XrpCurrencyAmount.ofDrops(10))
-      .sequence(UnsignedInteger.ONE)
-      .amount(XrpCurrencyAmount.ofDrops(10000))
-      .destination(Address.of("rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW"))
-      .destinationTag(UnsignedInteger.valueOf(23480))
-      .settleDelay(UnsignedInteger.valueOf(86400))
-      .publicKey("32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A")
-      .cancelAfter(UnsignedLong.valueOf(533171558))
-      .signingPublicKey(
-        PublicKey.fromBase16EncodedPublicKey("02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC")
-      )
-      .networkId(NetworkId.of(1024))
-      .build();
-
-    String json = "{" +
-      "  \"Account\": \"rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn\"," +
-      "  \"Fee\": \"10\"," +
-      "  \"Sequence\": 1," +
-      "  \"TransactionType\": \"PaymentChannelCreate\"," +
-      "  \"Amount\": \"10000\"," +
-      "  \"Destination\": \"rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW\"," +
-      "  \"SettleDelay\": 86400," +
-      "  \"PublicKey\": \"32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A\"," +
-      "  \"CancelAfter\": 533171558," +
-      "  \"DestinationTag\": 23480," +
-      "  \"SigningPubKey\" : \"02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC\"," +
-      "  \"NetworkID\": 1024," +
-      "  \"SourceTag\": 11747" +
-      "}";
-
-    assertCanSerializeAndDeserialize(create, json);
-  }
-
-  @Test
-  public void testPaymentChannelCreateJsonWithUnsetFlags() throws JsonProcessingException, JSONException {
-    PaymentChannelCreate create = PaymentChannelCreate.builder()
-      .account(Address.of("rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn"))
-      .sourceTag(UnsignedInteger.valueOf(11747))
-      .fee(XrpCurrencyAmount.ofDrops(10))
-      .sequence(UnsignedInteger.ONE)
-      .amount(XrpCurrencyAmount.ofDrops(10000))
-      .destination(Address.of("rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW"))
-      .destinationTag(UnsignedInteger.valueOf(23480))
-      .settleDelay(UnsignedInteger.valueOf(86400))
-      .publicKey("32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A")
-      .cancelAfter(UnsignedLong.valueOf(533171558))
-      .signingPublicKey(
-        PublicKey.fromBase16EncodedPublicKey("02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC")
-      )
-      .flags(TransactionFlags.UNSET)
-      .build();
-
-    String json = "{" +
-      "  \"Account\": \"rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn\"," +
-      "  \"Fee\": \"10\"," +
-      "  \"Sequence\": 1," +
-      "  \"TransactionType\": \"PaymentChannelCreate\"," +
-      "  \"Amount\": \"10000\"," +
-      "  \"Destination\": \"rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW\"," +
-      "  \"SettleDelay\": 86400," +
-      "  \"PublicKey\": \"32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A\"," +
-      "  \"CancelAfter\": 533171558," +
-      "  \"DestinationTag\": 23480," +
-      "  \"Flags\": 0," +
-      "  \"SigningPubKey\" : \"02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC\"," +
-      "  \"SourceTag\": 11747" +
-      "}";
-
-    assertCanSerializeAndDeserialize(create, json);
-  }
-
-  @Test
-  public void testPaymentChannelCreateJsonWithNonZeroFlags() throws JsonProcessingException, JSONException {
-    PaymentChannelCreate create = PaymentChannelCreate.builder()
-      .account(Address.of("rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn"))
-      .sourceTag(UnsignedInteger.valueOf(11747))
-      .fee(XrpCurrencyAmount.ofDrops(10))
-      .sequence(UnsignedInteger.ONE)
-      .amount(XrpCurrencyAmount.ofDrops(10000))
-      .destination(Address.of("rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW"))
-      .destinationTag(UnsignedInteger.valueOf(23480))
-      .settleDelay(UnsignedInteger.valueOf(86400))
-      .publicKey("32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A")
-      .cancelAfter(UnsignedLong.valueOf(533171558))
-      .signingPublicKey(
-        PublicKey.fromBase16EncodedPublicKey("02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC")
-      )
-      .flags(TransactionFlags.FULLY_CANONICAL_SIG)
-      .build();
-
-    String json = String.format("{" +
-      "  \"Account\": \"rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn\"," +
-      "  \"Fee\": \"10\"," +
-      "  \"Sequence\": 1," +
-      "  \"TransactionType\": \"PaymentChannelCreate\"," +
-      "  \"Amount\": \"10000\"," +
-      "  \"Destination\": \"rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW\"," +
-      "  \"SettleDelay\": 86400," +
-      "  \"PublicKey\": \"32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A\"," +
-      "  \"CancelAfter\": 533171558," +
-      "  \"DestinationTag\": 23480," +
-      "  \"Flags\": %s," +
-      "  \"SigningPubKey\" : \"02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC\"," +
-      "  \"SourceTag\": 11747" +
-      "}", TransactionFlags.FULLY_CANONICAL_SIG.getValue());
-
-    assertCanSerializeAndDeserialize(create, json);
-  }
-
-  @Test
-  public void testPaymentChannelCreateJsonWithUnknownFields() throws JsonProcessingException, JSONException {
-    PaymentChannelCreate create = PaymentChannelCreate.builder()
-      .account(Address.of("rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn"))
-      .sourceTag(UnsignedInteger.valueOf(11747))
-      .fee(XrpCurrencyAmount.ofDrops(10))
-      .sequence(UnsignedInteger.ONE)
-      .amount(XrpCurrencyAmount.ofDrops(10000))
-      .destination(Address.of("rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW"))
-      .destinationTag(UnsignedInteger.valueOf(23480))
-      .settleDelay(UnsignedInteger.valueOf(86400))
-      .publicKey("32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A")
-      .cancelAfter(UnsignedLong.valueOf(533171558))
-      .signingPublicKey(
-        PublicKey.fromBase16EncodedPublicKey("02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC")
-      )
-      .networkId(NetworkId.of(1024))
-      .putUnknownFields("Foo", "Bar")
-      .build();
-
-    String json = "{" +
-      "  \"Foo\" : \"Bar\"," +
-      "  \"Account\": \"rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn\"," +
-      "  \"Fee\": \"10\"," +
-      "  \"Sequence\": 1," +
-      "  \"TransactionType\": \"PaymentChannelCreate\"," +
-      "  \"Amount\": \"10000\"," +
-      "  \"Destination\": \"rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW\"," +
-      "  \"SettleDelay\": 86400," +
-      "  \"PublicKey\": \"32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A\"," +
-      "  \"CancelAfter\": 533171558," +
-      "  \"DestinationTag\": 23480," +
-      "  \"SigningPubKey\" : \"02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC\"," +
-      "  \"NetworkID\": 1024," +
-      "  \"SourceTag\": 11747" +
-      "}";
-
-    assertCanSerializeAndDeserialize(create, json);
-  }
-
-  @Test
-  public void testPaymentChannelClaimJson() throws JsonProcessingException, JSONException {
-    PaymentChannelClaim claim = PaymentChannelClaim.builder()
-      .account(Address.of("rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW"))
-      .fee(XrpCurrencyAmount.ofDrops(10))
-      .sequence(UnsignedInteger.ONE)
-      .channel(Hash256.of("C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198"))
-      .balance(XrpCurrencyAmount.ofDrops(1000000))
-      .amount(XrpCurrencyAmount.ofDrops(1000000))
-      .signature("30440220718D264EF05CAED7C781FF6DE298DCAC68D002562C9BF3A07C1E721B420C0DAB02203A5A4779E" +
-        "F4D2CCC7BC3EF886676D803A9981B928D3B8ACA483B80ECA3CD7B9B")
-      .publicKey("32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A")
-      .signingPublicKey(
-        PublicKey.fromBase16EncodedPublicKey("02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC")
-      )
-      .credentialIds(
-        Collections.singletonList(
-          Hash256.of("02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37")))
-      .networkId(NetworkId.of(1024))
-      .build();
-
-    String json = "{" +
-      "  \"Account\": \"rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW\"," +
-      "  \"Fee\": \"10\"," +
-      "  \"Sequence\": 1," +
-      "  \"TransactionType\": \"PaymentChannelClaim\"," +
-      "  \"Channel\": \"C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198\"," +
-      "  \"Balance\": \"1000000\"," +
-      "  \"Amount\": \"1000000\"," +
-      "  \"NetworkID\": 1024," +
-      "  \"Signature\": \"30440220718D264EF05CAED7C781FF6DE298DCAC68D002562C9BF3A07C1E721B420C0DAB02203A5A4" +
-      "779EF4D2CCC7BC3EF886676D803A9981B928D3B8ACA483B80ECA3CD7B9B\"," +
-      "    \"SigningPubKey\" : \"02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC\"," +
-      "    \"CredentialIDs\": [02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37]," +
-      "  \"PublicKey\": \"32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A\"" +
-      "}";
-
-    assertCanSerializeAndDeserialize(claim, json);
-  }
-
-  @Test
-  public void testPaymentChannelClaimJsonWithFlags() throws JsonProcessingException, JSONException {
-    PaymentChannelClaim claim = PaymentChannelClaim.builder()
-      .account(Address.of("rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW"))
-      .fee(XrpCurrencyAmount.ofDrops(10))
-      .sequence(UnsignedInteger.ONE)
-      .flags(PaymentChannelClaimFlags.builder().tfClose(true).build())
-      .channel(Hash256.of("C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198"))
-      .balance(XrpCurrencyAmount.ofDrops(1000000))
-      .amount(XrpCurrencyAmount.ofDrops(1000000))
-      .signature("30440220718D264EF05CAED7C781FF6DE298DCAC68D002562C9BF3A07C1E721B420C0DAB02203A5A4779E" +
-        "F4D2CCC7BC3EF886676D803A9981B928D3B8ACA483B80ECA3CD7B9B")
-      .publicKey("32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A")
-      .signingPublicKey(
-        PublicKey.fromBase16EncodedPublicKey("02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC")
-      )
-      .build();
-
-    String json = "{" +
-      "  \"Account\": \"rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW\"," +
-      "  \"Fee\": \"10\"," +
-      "  \"Sequence\": 1," +
-      "  \"Flags\": 2147614720," +
-      "  \"TransactionType\": \"PaymentChannelClaim\"," +
-      "  \"Channel\": \"C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198\"," +
-      "  \"Balance\": \"1000000\"," +
-      "  \"Amount\": \"1000000\"," +
-      "  \"Signature\": \"30440220718D264EF05CAED7C781FF6DE298DCAC68D002562C9BF3A07C1E721B420C0DAB02203A5A4" +
-      "779EF4D2CCC7BC3EF886676D803A9981B928D3B8ACA483B80ECA3CD7B9B\"," +
-      "    \"SigningPubKey\" : \"02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC\"," +
-      "  \"PublicKey\": \"32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A\"" +
-      "}";
-
-    assertCanSerializeAndDeserialize(claim, json);
-  }
-
-  @Test
-  public void testPaymentChannelClaimJsonWithUnknownFields() throws JsonProcessingException, JSONException {
-    PaymentChannelClaim claim = PaymentChannelClaim.builder()
-      .account(Address.of("rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW"))
-      .fee(XrpCurrencyAmount.ofDrops(10))
-      .sequence(UnsignedInteger.ONE)
-      .channel(Hash256.of("C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198"))
-      .balance(XrpCurrencyAmount.ofDrops(1000000))
-      .amount(XrpCurrencyAmount.ofDrops(1000000))
-      .signature("30440220718D264EF05CAED7C781FF6DE298DCAC68D002562C9BF3A07C1E721B420C0DAB02203A5A4779E" +
-        "F4D2CCC7BC3EF886676D803A9981B928D3B8ACA483B80ECA3CD7B9B")
-      .publicKey("32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A")
-      .signingPublicKey(
-        PublicKey.fromBase16EncodedPublicKey("02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC")
-      )
-      .networkId(NetworkId.of(1024))
-      .putUnknownFields("Foo", "Bar")
-      .build();
-
-    String json = "{" +
-      "  \"Foo\" : \"Bar\"," +
-      "  \"Account\": \"rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW\"," +
-      "  \"Fee\": \"10\"," +
-      "  \"Sequence\": 1," +
-      "  \"TransactionType\": \"PaymentChannelClaim\"," +
-      "  \"Channel\": \"C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198\"," +
-      "  \"Balance\": \"1000000\"," +
-      "  \"Amount\": \"1000000\"," +
-      "  \"NetworkID\": 1024," +
-      "  \"Signature\": \"30440220718D264EF05CAED7C781FF6DE298DCAC68D002562C9BF3A07C1E721B420C0DAB02203A5A4" +
-      "779EF4D2CCC7BC3EF886676D803A9981B928D3B8ACA483B80ECA3CD7B9B\"," +
-      "    \"SigningPubKey\" : \"02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC\"," +
-      "  \"PublicKey\": \"32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A\"" +
-      "}";
-
-    assertCanSerializeAndDeserialize(claim, json);
-  }
-
-  @Test
-  public void testPaymentChannelFundJson() throws JsonProcessingException, JSONException {
-    PaymentChannelFund fund = PaymentChannelFund.builder()
-      .account(Address.of("rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn"))
-      .fee(XrpCurrencyAmount.ofDrops(10))
-      .sequence(UnsignedInteger.ONE)
-      .channel(Hash256.of("C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198"))
-      .amount(XrpCurrencyAmount.ofDrops(200000))
-      .expiration(UnsignedLong.valueOf(543171558))
-      .signingPublicKey(
-        PublicKey.fromBase16EncodedPublicKey("02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC")
-      )
-      .networkId(NetworkId.of(1024))
-      .build();
-
-    String json = "{" +
-      "  \"Account\": \"rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn\"," +
-      "  \"Fee\": \"10\"," +
-      "  \"Sequence\": 1," +
-      "  \"TransactionType\": \"PaymentChannelFund\"," +
-      "  \"Channel\": \"C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198\"," +
-      "  \"Amount\": \"200000\"," +
-      "  \"NetworkID\": 1024," +
-      "  \"SigningPubKey\" : \"02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC\"," +
-      "  \"Expiration\": 543171558" +
-      "}";
-
-    assertCanSerializeAndDeserialize(fund, json);
-  }
-
-  @Test
-  public void testPaymentChannelFundJsonWithUnsetFlags() throws JsonProcessingException, JSONException {
-    PaymentChannelFund fund = PaymentChannelFund.builder()
-      .account(Address.of("rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn"))
-      .fee(XrpCurrencyAmount.ofDrops(10))
-      .sequence(UnsignedInteger.ONE)
-      .channel(Hash256.of("C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198"))
-      .amount(XrpCurrencyAmount.ofDrops(200000))
-      .expiration(UnsignedLong.valueOf(543171558))
-      .signingPublicKey(
-        PublicKey.fromBase16EncodedPublicKey("02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC")
-      )
-      .flags(TransactionFlags.UNSET)
-      .build();
-
-    String json = "{" +
-      "  \"Account\": \"rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn\"," +
-      "  \"Fee\": \"10\"," +
-      "  \"Sequence\": 1," +
-      "  \"TransactionType\": \"PaymentChannelFund\"," +
-      "  \"Channel\": \"C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198\"," +
-      "  \"Amount\": \"200000\"," +
-      "  \"Flags\": 0," +
-      "  \"SigningPubKey\" : \"02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC\"," +
-      "  \"Expiration\": 543171558" +
-      "}";
-
-    assertCanSerializeAndDeserialize(fund, json);
-  }
-
-  @Test
-  public void testPaymentChannelFundJsonWithNonZeroFlags() throws JsonProcessingException, JSONException {
-    PaymentChannelFund fund = PaymentChannelFund.builder()
-      .account(Address.of("rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn"))
-      .fee(XrpCurrencyAmount.ofDrops(10))
-      .sequence(UnsignedInteger.ONE)
-      .channel(Hash256.of("C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198"))
-      .amount(XrpCurrencyAmount.ofDrops(200000))
-      .expiration(UnsignedLong.valueOf(543171558))
-      .signingPublicKey(
-        PublicKey.fromBase16EncodedPublicKey("02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC")
-      )
-      .flags(TransactionFlags.FULLY_CANONICAL_SIG)
-      .build();
-
-    String json = String.format("{" +
-      "  \"Account\": \"rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn\"," +
-      "  \"Fee\": \"10\"," +
-      "  \"Sequence\": 1," +
-      "  \"TransactionType\": \"PaymentChannelFund\"," +
-      "  \"Channel\": \"C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198\"," +
-      "  \"Amount\": \"200000\"," +
-      "  \"Flags\": %s," +
-      "  \"SigningPubKey\" : \"02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC\"," +
-      "  \"Expiration\": 543171558" +
-      "}", TransactionFlags.FULLY_CANONICAL_SIG.getValue());
-
-    assertCanSerializeAndDeserialize(fund, json);
-  }
-
-  @Test
-  public void testPaymentChannelFundJsonWithUnknownFields() throws JsonProcessingException, JSONException {
-    PaymentChannelFund fund = PaymentChannelFund.builder()
-      .account(Address.of("rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn"))
-      .fee(XrpCurrencyAmount.ofDrops(10))
-      .sequence(UnsignedInteger.ONE)
-      .channel(Hash256.of("C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198"))
-      .amount(XrpCurrencyAmount.ofDrops(200000))
-      .expiration(UnsignedLong.valueOf(543171558))
-      .signingPublicKey(
-        PublicKey.fromBase16EncodedPublicKey("02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC")
-      )
-      .networkId(NetworkId.of(1024))
-      .putUnknownFields("Foo", "Bar")
-      .build();
-
-    String json = "{" +
-      "  \"Foo\" : \"Bar\"," +
-      "  \"Account\": \"rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn\"," +
-      "  \"Fee\": \"10\"," +
-      "  \"Sequence\": 1," +
-      "  \"TransactionType\": \"PaymentChannelFund\"," +
-      "  \"Channel\": \"C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198\"," +
-      "  \"Amount\": \"200000\"," +
-      "  \"NetworkID\": 1024," +
-      "  \"SigningPubKey\" : \"02356E89059A75438887F9FEE2056A2890DB82A68353BE9C0C0C8F89C0018B37FC\"," +
-      "  \"Expiration\": 543171558" +
-      "}";
-
-    assertCanSerializeAndDeserialize(fund, json);
-  }
-}
